requirements.txt

streamlit
faiss-cpu
sentence-transformers
pdfplumber
requests